+++ Lib/fontbakery/profiles/googlefonts.py
@@ -64,6 +64,7 @@ DESCRIPTION_CHECKS = [
    'com.google.fonts/check/description/min_length',
    'com.google.fonts/check/description/max_length',
    'com.google.fonts/check/description/git_url',
+   'com.google.fonts/check/description/variable_font'
 ]
 
 FAMILY_CHECKS = [
@@ -263,6 +264,30 @@ def com_google_fonts_check_description_git_url(description):
     yield FAIL, ("Please host your font project on a public Git repo"
                  " (such as GitHub or GitLab) and place a link"
                  " in the DESCRIPTION.en_us.html file.")
+
+
+@check(
+  id = 'com.google.fonts/check/description/variable_font',
+  conditions = ['is_variable_font',
+                'description'],
+  rationale = """
+    Families with variable fonts do not always mention that
+    in their descriptions.
+
+    Therefore, this check ensures that a standard boilerplate
+    sentence is present in the DESCRIPTION files for all
+    those families which are available as variable fonts.
+  """
+)
+def com_google_fonts_check_description_variable_font(description):
+  """Does DESCRIPTION file mention when a family
+     is available as variable font?"""
+  if "variable font" not in description.lower():
+    yield FAIL, ("Please include the following sentence"
+                 " in the DESCRIPTION.en-us.html file:\n\n"
+                 "This family is available as a variable font.")
+  else:
+    yield PASS, "Looks good!"
 
 
 @check(